requirements.txt
# ============================================================
# CL-BEDS Backend Requirements
# Python 3.11 | FastAPI | Supabase PostgreSQL | Render
# ============================================================

# --- Web framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# --- Config & Validation ---
pydantic==2.10.3
pydantic-settings==2.6.1

# --- Database ---
sqlalchemy==2.0.36
asyncpg==0.30.0

# --- Auth ---
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# --- HTTP client ---
httpx==0.28.1

# --- ML / AI ---
torch==2.5.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
numpy==2.1.3
transformers==4.47.1
tokenizers==0.21.0

# --- Settings ---
python-dotenv==1.0.1
